v1 of credential

+++ internal/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID              int    `json:"id"`
+	ID              int64  `json:"id"`
 	IsSuperuser     bool   `json:"is_superuser"`
 	IsSystemAuditor bool   `json:"is_system_auditor"`
 	Username        string `json:"username"`
